requirements.txt
# Data manipulation & analysis
pandas==2.1.4
numpy==1.26.4

# Database connectivity (PostgreSQL)
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Fake data generation
faker==22.2.0

# Configuration file parsing (YAML)
pyyaml==6.0.1

# Environment variable management
python-dotenv==1.0.1

# Testing & coverage
pytest==8.0.0
pytest-cov==4.1.0

# Task scheduling / orchestration (lightweight)
schedule==1.2.1
